Moved login page to root directory
Moved login page to root directory to be the initial landing page
